Re-arrange code into smaller files based on functionality Change default CSI Spec in the default config to 0.2 Add support for snapshots Add support to handle variations in sidecars based on CSI Spec version
