+++ deploy/install.yaml
@@ -206,6 +206,34 @@ rules:
   - watch
   - update
   - delete
+- apiGroups: 
+  - snapshot.storage.k8s.io
+  resources: 
+  - volumesnapshotclasses
+  verbs: 
+  - get
+  - list
+  - watch
+- apiGroups:
+  - snapshot.storage.k8s.io
+  resources: 
+  - volumesnapshotcontents
+  verbs:
+  - create
+  - get
+  - list
+  - watch
+  - update
+  - delete
+- apiGroups:
+  - snapshot.storage.k8s.io
+  resources: 
+  - volumesnapshots
+  verbs:
+  - get
+  - list
+  - watch
+  - update
 
 ---
 kind: ClusterRoleBinding
